Insert Flatiron Solution. Note problems below:
With my solution (in tictactoe.js), using mocha.run(), all but two tests
pass (the last two tests of gameplay). Yet, with my solution, when I type
"learn" in the terminal, 15 tests are not passing.  Discrepancy is not
clear to me. It is also not clear what the last two tests of gameplay
are mean to get at.